% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scvelo_utility_functions.R
\name{addSUmatrices}
\alias{addSUmatrices}
\title{Add spliced and unspliced matrices from a Loom file to a Seurat object}
\usage{
addSUmatrices(X, loomFile)
}
\arguments{
\item{X}{Seurat object to which spliced and unspliced matrices will be added}

\item{loomFile}{path to the Loom file containing spliced and unspliced matrices}
}
\value{
Seurat object with spliced and unspliced matrices added as new assays
}
\description{
This function takes a Seurat object and a Loom file as inputs. It reads spliced and 
unspliced matrices from the Loom file and adds them as new assays to the Seurat object. 
The function first gets standardized barcodes for the Loom file and renames the cells in 
the Seurat object to match the barcodes. It then identifies the shared barcodes between 
the two datasets, filters the Seurat object to match the barcodes in the Loom file, and adds 
the spliced and unspliced matrices as new assays.
}
