% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat_functions.R
\name{aggregate_visu}
\alias{aggregate_visu}
\title{Take as input a CellChat object and graph the aggregated
cell-cell communication network.}
\usage{
aggregate_visu(X, condition)
}
\arguments{
\item{X}{a CellChat object.}

\item{condition}{the condition of the object used for naming files.}
}
\description{
The CellChat object needs to have centrality scores calculated.
}
