% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat_functions.R
\name{align_cell_labels}
\alias{align_cell_labels}
\title{Take as inputs two CellChat objects with different cell type labels.
Prepare the two objects for merging by aligning cell type labels,
net values, and netP values.}
\usage{
align_cell_labels(X1, X2)
}
\arguments{
\item{X1}{a CellChat object.}

\item{X2}{a CellChat object.}
}
\value{
CellChat objects with aligned labels.
}
\description{
Take as inputs two CellChat objects with different cell type labels.
Prepare the two objects for merging by aligning cell type labels,
net values, and netP values.
}
