% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-extended.R
\name{check_annotation}
\alias{check_annotation}
\title{Check annotation with marker panels}
\usage{
check_annotation(obj, outdir = NULL)
}
\arguments{
\item{obj}{A \code{scdown_obj}.}

\item{outdir}{Optional output directory.}
}
\value{
A list with annotation-check tables and plots.
}
\description{
Check annotation with marker panels
}
