% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{cleanSeuratMeta}
\alias{cleanSeuratMeta}
\title{Cleanup unsafe characters in the Seurat meta.data and Idents.}
\usage{
cleanSeuratMeta(seurat_obj)
}
\arguments{
\item{seurat_obj}{Seurat object}
}
\value{
a Seurat object.
}
\description{
This function converts unsafe characters in a file name to a underscore.
}
