% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{collapse_lineage}
\alias{collapse_lineage}
\title{Collapse fine labels into major lineages}
\usage{
collapse_lineage(obj, mapping = NULL)
}
\arguments{
\item{obj}{A \code{scdown_obj}.}

\item{mapping}{Optional data frame with \code{pattern} and \code{lineage}.}
}
\value{
The updated \code{scdown_obj}.
}
\description{
Collapse fine labels into major lineages
}
