% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat_functions.R
\name{compareCellComVisu}
\alias{compareCellComVisu}
\title{Do the workflow for cell-cell communication analysis on two
biological conditions and output relevant visualizations.}
\usage{
compareCellComVisu(X, object_list, cond_in_compare, pathways_to_compare)
}
\arguments{
\item{X}{a merged CellChat object with two biological conditions.}

\item{object_list}{a list of CellChat object before the merge.}

\item{cond_in_compare}{a vector of condition names in comparison.}

\item{pathways_to_compare}{a vector of pathway names.}
}
\value{
X a CellChat object after pairwise comparison workflow.
}
\description{
Do the workflow for cell-cell communication analysis on two
biological conditions and output relevant visualizations.
}
