% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat_functions.R
\name{create_dir_cellchat}
\alias{create_dir_cellchat}
\title{Create result directories for CellChat.}
\usage{
create_dir_cellchat(dir_cellchat)
}
\arguments{
\item{dir_cellchat}{Folder path for CellChat results.}
}
\description{
Takes the folder path to create folers to store the CellChat results 
includeing figures, tables and rds files
}
