% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scProportionTest_functions.R
\name{create_dir_scproportion}
\alias{create_dir_scproportion}
\title{Create result directories for scProportionTest figures}
\usage{
create_dir_scproportion(dir_scproportion)
}
\arguments{
\item{dir_scproportion}{Folder path for scProportionTest figures}
}
\description{
Create result directories for scProportionTest figures
}
