% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{default_lineage_mapping}
\alias{default_lineage_mapping}
\title{Default lineage mapping}
\usage{
default_lineage_mapping()
}
\value{
A data frame with pattern and lineage.
}
\description{
Default lineage mapping
}
