% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat_functions.R
\name{differential_ligand_receptor}
\alias{differential_ligand_receptor}
\title{Take as input a CellChat object and output the differential
ligand/receptor pair analysis result.}
\usage{
differential_ligand_receptor(X, cond_in_compare)
}
\arguments{
\item{X}{a merged CellChat object with two biological conditions.}

\item{cond_in_compare}{a vector of condition names in comparison.}
}
\value{
X a CellChat object with differential LR pair results.
}
\description{
Take as input a CellChat object and output the differential
ligand/receptor pair analysis result.
}
