% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat_functions.R
\name{doCellCom}
\alias{doCellCom}
\title{Do cell-cell communication analysis using cellchat}
\usage{
doCellCom(X, species)
}
\arguments{
\item{X}{a Seurat object.}

\item{species}{a species name.}
}
\value{
A CellChat object with cell-cell communication analysis results.
}
\description{
Takes as input a Seurat object with cell-type labels as identity of the cells.
This Seurat object should have identities populated at Idents(X) and counts
normalized in X@assays$RNA@data.
}
