% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat_functions.R
\name{doCellComVisu}
\alias{doCellComVisu}
\title{Takes as input a CellChat object with communication analysis results
and a vector of pathway names to show. Call aggregate_visu, pathway_visu,
to graph all visualizations.}
\usage{
doCellComVisu(X, Y, pathways_to_show, condition)
}
\arguments{
\item{X}{a CellChat object.}

\item{pathways_to_show}{a vector of pathway names.}

\item{condition}{the condition of the object.}
}
\description{
Takes as input a CellChat object with communication analysis results
and a vector of pathway names to show. Call aggregate_visu, pathway_visu,
to graph all visualizations.
}
