% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scvelo_utility_functions.R
\name{doVelocity}
\alias{doVelocity}
\title{Compute single-cell velocity using scvelo package}
\usage{
doVelocity(X, mode = "stochastic")
}
\arguments{
\item{X}{Seurat object containing the counts matrices.}

\item{mode}{Character mode specifying the type of velocity computation to use. Available modes are "steady_state" (original), "deterministic", "stochastic" (fastest:recommended), "dynamical".}
}
\value{
A SingleCellExperiment object with the information required to generate the plots.
}
\description{
The function computes the single-cell velocity using the scvelo package.
The function takes in total, spliced and unspliced counts from a Seurat 
object and uses them to calculate single-cell velocity. The function 
returns a SingleCellExperiment object with the information required to 
generate the plots.
}
