% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore_composition}
\alias{explore_composition}
\title{Explore composition summaries}
\usage{
explore_composition(obj, outdir = NULL)
}
\arguments{
\item{obj}{A \code{scdown_obj}.}

\item{outdir}{Optional output directory.}
}
\value{
A list with composition tables and plot paths.
}
\description{
Explore composition summaries
}
