% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore_signature}
\alias{explore_signature}
\title{Explore signature scores}
\usage{
explore_signature(obj, signatures, signature_sets = NULL, outdir = NULL)
}
\arguments{
\item{obj}{A \code{scdown_obj}.}

\item{signatures}{Character vector of signature names or a named list of signatures.}

\item{signature_sets}{Optional named list of custom signatures to merge with built-ins.}

\item{outdir}{Optional output directory.}
}
\value{
A list with signature tables and plots.
}
\description{
Explore signature scores
}
