% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scProportionTest_functions.R
\name{generate_figure}
\alias{generate_figure}
\title{Generate plot for each comparision}
\usage{
generate_figure(
  prop_test.i,
  output.format,
  comparisons_condition,
  dir_scproportion,
  cluster_col,
  i
)
}
\arguments{
\item{prop_test.i}{scProportion object}

\item{output.format}{The format of output figure}

\item{i}{index of the comparision condition}
}
\value{
NULL saves comparison figures in the specified directory.
}
\description{
Generate plot for each comparision
}
