% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-helpers.R
\name{get_signature}
\alias{get_signature}
\title{Get one built-in signature}
\usage{
get_signature(name)
}
\arguments{
\item{name}{Signature name.}
}
\value{
Character vector of genes.
}
\description{
Get one built-in signature
}
