% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAvelocity_functions.R
\name{getVectorField}
\alias{getVectorField}
\title{Compute vector field for plotting velocity arrows using scvelo package}
\usage{
getVectorField(
  X,
  scVeloOutput,
  reduction = "umap",
  dims = 1:2,
  resolution = 50
)
}
\arguments{
\item{X}{Seurat object containing the counts matrices.}

\item{scVeloOutput}{scVeloOutput object returned by doVelocity function.}

\item{reduction}{Character specifying the dimensionality reduction method to use for plotting. Default is "umap".}

\item{dims}{Numeric vector specifying the dimensions to be plotted. Default is 1:2.}

\item{resolution}{Numeric value specifying the resolution of the grid to use. Default is 50.}
}
\value{
A 4-column data.frame containing the x and y coordinates of the start and end points of the velocity arrows.
}
\description{
The function computes the vector field for plotting velocity arrows 
using the scvelo package. The function takes in a Seurat object 
containing the counts matrices, the scVeloOutput object returned by 
doVelocity function, the reduction method used for plotting (default: umap), 
dimensions to be plotted (default: 1:2) and the resolution of the grid (default: 50). 
The function returns a 4-column data.frame that can be used with 
geom_segment to plot velocity arrows.
}
