% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAvelocity_functions.R
\name{getVelocityPseudotime}
\alias{getVelocityPseudotime}
\title{Transfer computed pseudotime values using scvelo package to a Seurat object}
\usage{
getVelocityPseudotime(X, scVeloOutput)
}
\arguments{
\item{X}{Seurat object containing the counts matrices.}

\item{scVeloOutput}{scVeloOutput object returned by doVelocity function.}
}
\value{
A Seurat object with two additional columns containing the computed pseudotime values and the velocity confidence values.
}
\description{
The function transfers the computed pseudotime values from the scvelo package
to the Seurat object. The function takes in a Seurat object containing the 
counts matrices and a scVeloOutput object returned by doVelocity function. 
The function returns the Seurat object with two additional columns in the metadata
containing the computed pseudotime values and the velocity confidence values.
}
