% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-helpers.R
\name{immune_marker_panels}
\alias{immune_marker_panels}
\title{Built-in immune marker panels}
\usage{
immune_marker_panels()
}
\value{
A named list of character vectors.
}
\description{
Built-in immune marker panels
}
