% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat_functions.R
\name{information_flow}
\alias{information_flow}
\title{Take as input a CellChat object and output the information
flow comparison results.}
\usage{
information_flow(X, object_list, cond_in_compare)
}
\arguments{
\item{X}{a merged CellChat object with two biological conditions.}

\item{object_list}{a list of CellChat objects before the merge.}

\item{cond_in_compare}{a vector of condition names in comparison.}
}
\description{
Take as input a CellChat object and output the information
flow comparison results.
}
