% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{list_signatures}
\alias{list_signatures}
\title{List available signatures}
\usage{
list_signatures(obj = NULL)
}
\arguments{
\item{obj}{Optional \code{scdown_obj} with custom signatures.}
}
\value{
A character vector.
}
\description{
List available signatures
}
