% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat_functions.R
\name{manifold_learning}
\alias{manifold_learning}
\title{Take as input a CellChat object and quantify the similarity between
all significant signaling pathways and then group them based on their
cellular communication network similarity.}
\usage{
manifold_learning(X, cond_in_compare)
}
\arguments{
\item{X}{a CellChat object.}

\item{cond_in_compare}{a vector of condition names in comparison.}
}
\value{
X a CellChat object with manifold learning results.
}
\description{
Need to use conda env to run python "umap-learn" package
}
