% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat_functions.R
\name{network_comparison}
\alias{network_comparison}
\title{Take as input a merged CellChat object and a list of CellChat object
prior to the merge, and output the general comparison results such as
number of interactions and aggregated interaction strength.}
\usage{
network_comparison(X, object_list, cond_in_compare)
}
\arguments{
\item{X}{a merged CellChat object with two biological conditions.}

\item{object_list}{a list of CellChat objects before the merge.}

\item{cond_in_compare}{a vector of condition names in comparison.}
}
\description{
Take as input a merged CellChat object and a list of CellChat object
prior to the merge, and output the general comparison results such as
number of interactions and aggregated interaction strength.
}
