% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat_functions.R
\name{pathway_visu}
\alias{pathway_visu}
\title{Take as input a CellChat object and a pathway in which one
wants to focus on. Graph the communication network for that
specific pathway.}
\usage{
pathway_visu(X, Y, pathway, condition)
}
\arguments{
\item{X}{a CellChat object.}

\item{pathway}{a signaling pathway in interest.}

\item{condition}{the condition of the object used for naming files.}
}
\description{
Could add one more parameter to choose specific figure layouts.
The CellChat object needs to have centrality scores calculated.
}
