% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-basic.R
\name{plot_gene}
\alias{plot_gene}
\title{Plot user-selected genes}
\usage{
plot_gene(obj, genes, outdir = NULL)
}
\arguments{
\item{obj}{A \code{scdown_obj}.}

\item{genes}{Character vector of genes to visualize.}

\item{outdir}{Optional output directory.}
}
\value{
A list with gene tables and plot paths.
}
\description{
Plot user-selected genes
}
