% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream.R
\name{plot_map}
\alias{plot_map}
\title{Plot cluster or annotation maps}
\usage{
plot_map(obj, outdir = NULL)
}
\arguments{
\item{obj}{A \code{scdown_obj}.}

\item{outdir}{Optional output directory.}
}
\value{
A list with tables and plot paths.
}
\description{
Plot cluster or annotation maps
}
