% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scvelo_utility_functions.R
\name{plotPseudotime}
\alias{plotPseudotime}
\title{Plot the pseudotime on lower-dimensional embeddings.}
\usage{
plotPseudotime(X, color_scale = c("darkblue", "yellow"))
}
\arguments{
\item{X}{a Seurat object outputted by getVelocityPseudotime() function.}

\item{color_scale}{two colors to use to define color gradient when plotting pseudotime.}
}
\description{
This function takes in a Seurat object outputted by getVelocityPseudotime() function, which
contains pseudotime values in its metadata. Lower-dimensional space is plotted using geom_point
and pseudotime values are used as scales for coloring.
}
