% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scvelo_utility_functions.R
\name{plotVectorField}
\alias{plotVectorField}
\title{Plot the vector field on lower-dimensional embeddings.}
\usage{
plotVectorField(
  X,
  tpVF,
  time_point = NULL,
  time_point_column = NULL,
  color_scale = NULL,
  name_by = NULL,
  grid_res = NULL,
  arrow_size = 0.5,
  vector_width = 0.5
)
}
\arguments{
\item{X}{a Seurat object with dimension reduction coordinates.}

\item{tpVF}{the 4-column dataframe outputted by getVectorField() with start and end point of each velocity vector.}

\item{time_point}{a character vector of timepoints in the datasets specifying cells from which time should be plotted with velocity arrows.}

\item{time_point_column}{a character string specifying name of the metadata that has timepoint information.}

\item{color_scale}{a character vector of colors to be used in plotting.}

\item{name_by}{a character string specifying which metadata in X is the color scale named by.}

\item{grid_res}{numeric value specifying the resolution of the grid, purely for figure naming purpose.}

\item{arrow_size}{a float or integer specifying the velocity vector arrow head size.}

\item{vector_width}{a float or integer specifying the velocity vector line width.}
}
\description{
This function takes in a Seurat object with dimensionality reduction and the four-column
data frame outputted by getVectorField() function, which contains velocity vector start
and end coordinates. Then the function uses geom_point to plot the lower-dimensional space
and geom_segment to overlay the velocity vectors.
}
