% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat_functions.R
\name{reorder_ident}
\alias{reorder_ident}
\title{Reorder the cell identities and net calculation results based on given input levels.
https://github.com/sqjin/CellChat/issues/149#issuecomment-788102884}
\usage{
reorder_ident(object, ident.use, levels)
}
\arguments{
\item{object}{a CellChat object}

\item{ident.use}{the name of the variable in object.meta}

\item{levels}{set the levels of factor}
}
\value{
a CellChat object with updated cell type and net order.
}
\description{
Reorder the cell identities and net calculation results based on given input levels.
https://github.com/sqjin/CellChat/issues/149#issuecomment-788102884
}
