% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cellchatV2.R
\name{run_cellchatV2}
\alias{run_cellchatV2}
\title{Run CellChat}
\usage{
run_cellchatV2(
  seurat_obj,
  species,
  annotation_column,
  annotation_selected = NULL,
  sample_column = NULL,
  group_column = NULL,
  group_cmp = NULL,
  top_n_pathways = 10,
  output_format = "png",
  n_jobs = 4,
  output_dir = "."
)
}
\arguments{
\item{seurat_obj}{Seurat object containing the UMI counts and metadata for the single-cell data.}

\item{species}{The species of the data, either 'human' or 'mouse'.}

\item{annotation_column}{Name of the metadata column in the Seurat object that contains cell annotations.}

\item{annotation_selected}{A vector of cell annotations of interest for running the CellChat analysis. If not provided, all cell types will be considered.}

\item{sample_column}{Name of the metadata column in the Seurat object that contains sample information.}

\item{group_column}{Name of the metadata column in the Seurat object that defines conditions or groups.}

\item{group_cmp}{A list of pairwise condition comparisons for differential CellChat analysis (e.g., comparing different groups or conditions).}

\item{top_n_pathways}{The number of top signaling pathways to analyze and visualize.}

\item{output_format}{Format of output figure: "png" or "pdf" (default: "png")}

\item{n_jobs}{A numeric variable to set the number of cores to be used}

\item{output_dir}{Path to the folder where CellChat results will be saved (figures, tables, and RDS files).}
}
\description{
This function performs CellChat analysis on a Seurat object to investigate cell-to-cell communication.
It generates figures, tables, and RDS files as output. The function allows for the specification of
various parameters, including the directory for storing results, the metadata column for cell annotations,
and the cell annotations of interest. Additionally, users can provide species information, group metadata
for group comparisons, and define pairwise group for differential analysis. The function also supports
specifying the number of top pathways to analyze and visualize.
}
