% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_monocle3.R
\name{run_monocle3}
\alias{run_monocle3}
\title{Function to run the monocle3 pipeline}
\usage{
run_monocle3(
  seurat_obj,
  species,
  nDim = 30,
  groups = NULL,
  annotation_column = NULL,
  group_column = NULL,
  transferUMAP = TRUE,
  rootNode_method = "potency",
  rootNode = NULL,
  timePoint = NULL,
  timePoint_metadata = NULL,
  sample_column = NULL,
  celltype_groups = NULL,
  n_top_genes = 10,
  deg_method = "quasipoisson",
  metadata_deg_model = NULL,
  graph_test = FALSE,
  output_format = "png",
  n_jobs = 4,
  output_dir = "."
)
}
\arguments{
\item{seurat_obj}{Seurat object containing the scRNASeq data (Required)}

\item{species}{A character variable specifying the species (Required, either human or mouse)}

\item{nDim}{A numeric variable specifying the PCA dimensions to be used for processing}

\item{groups}{A character vector specifying the different conditions (groups)}

\item{annotation_column}{A character variable specifying the metdata column name of annotations. (Default: uses Idents)}

\item{group_column}{A character variable specifying the metdata column name containing conditions in the seurat object}

\item{transferUMAP}{A boolean variable specifying whether to transfer UMAP coordinates or not from Seurat object}

\item{rootNode_method}{A character variable specifying the method to identify root node (Either potency or rootNodes)}

\item{rootNode}{A character variable specifying the principal point on the trajectory, in the format of "Y_<node number>".}

\item{timePoint}{A character variable specifying the category inside a metadata to be used to automatically decide the root node.}

\item{timePoint_metadata}{A character variable specifying the metadata to be used to automatically decide the root node.}

\item{sample_column}{A character vector specifying the name of metadata column name with batch information to perform batch correction}

\item{celltype_groups}{A list of character vectors that represents a group of cell types to be subsetted and constructed trajectory by monocle}

\item{n_top_genes}{A numeric variable to set the number of top significant genes to visualize from DEG results}

\item{deg_method}{A character vector specifying the distribution used for regression method}

\item{metadata_deg_model}{A character vector specifying the model used for regression analysis, can be any term that exists in the metadata}

\item{graph_test}{A boolean variable to run graph auto-correlation to find differential genes along the trajectory}

\item{output_format}{Format of output figure: "png" or "pdf" (default: "png")}

\item{n_jobs}{A numeric variable to set the number of cores to be used}

\item{output_dir}{A character vector specifying the output directory}
}
\value{
A list of monocle cell dataset objects
}
\description{
This function runs the moncole3 workflow to infer the development trajectory of the cells using the
single-cell RNASeq data. In this function, users can infer the trajectory of the whole data as well as
a subset of data and also identify the differential genes across the trajectory as well as between the
conditions.
}
