% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sccomp.R
\name{run_sccomp}
\alias{run_sccomp}
\title{Function to run scComp pipeline}
\usage{
run_sccomp(
  seurat_obj,
  annotation_column,
  sample_column,
  group_column,
  output_dir = ".",
  output_format = "png",
  control_cond = NULL,
  significance_threshold = 0.05,
  additional_effects = NULL,
  verbose = TRUE,
  n_jobs = 1
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing count data and metadata.}

\item{annotation_column}{The name of the metadata column in `seurat_obj` containing cluster labels or cell type names.}

\item{sample_column}{The name of the metadata column in `seurat_obj` that contains sample identifiers.}

\item{group_column}{The name of the metadata column in `seurat_obj` that contains conditions for comparisons.}

\item{output_dir}{Directory path where output figures will be saved.}

\item{output_format}{Format of the output figure.}

\item{control_cond}{Control condition group label.}

\item{significance_threshold}{Threshold of significance for plotting purpose.}

\item{additional_effects}{Additional effects to be added to formula.}

\item{verbose}{Print the processing steps.}

\item{n_jobs}{A numeric variable to set the number of cores to be used.}
}
\value{
NULL Save comparison figures and results in the specified directory.
}
\description{
This function runs scComp to perform differential composition and variability analysis
and generate figures and table of results.
The user can optionally choose a control condition to compare all others against it.
}
