% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\name{scdown_example}
\alias{scdown_example}
\title{Example annotated scRNA-seq dataset}
\usage{
scdown_example()
}
\value{
A list with \code{expr}, \code{cells}, and \code{embedding}.
}
\description{
Example annotated scRNA-seq dataset
}
