% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{seuratToH5ad}
\alias{seuratToH5ad}
\title{Save a Seurat object to H5AD file.}
\usage{
seuratToH5ad(
  seurat_obj,
  loom_files = NULL,
  output_dir = ".",
  loom_file_subset_by = NULL,
  loom_file_subset_column = "orig.ident",
  annotation_column = "ID",
  groups = NULL,
  group_column = NULL
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{loom_files}{path and file names of Spliced and unspliced counts of the scRNA-seq data (Required if
not provided in `seurat_obj`). For multiple loom files, the file names in `loom_files` must correspond to
the values in the metadata column in the Seurat object specified by `loom_file_subset_column`}

\item{output_dir}{A character vector specifying the output directory}

\item{loom_file_subset_by}{A character variable specifying how the Seurat object should be subsetted in order
to match each indivdidual loom file name. This vector must correspond to a metadata column specified by
`loom_file_subset_column` in the Seurat object, and the order of `loom_file_subset_by` must match the order of
loom_files. If `loom_file_subset_by` is not provided (default `loom_file_subset_by <- NULL`), it will be
automatically extracted from file names in `loom_files` to ensure the correct order.
If there is only one loom file provided, this variable should be left blank: `loom_file_subset_by <- NULL`}

\item{loom_file_subset_column}{A string specifying the name of the metadata column in the Seurat object that
should be used for subsetting to match each of the loom files. This column must exist in the Seurat object
metadata. If there is only one loom file provided, this variable should be left blank:
`loom_file_subset_column <- NULL`; if there are multiple loom files provided, this variable needs to be provided.}

\item{annotation_column}{A character variable specifying which metadata column of the h5ad object contains
cell type annotations.}

\item{groups}{A list of character vectors representing groups of conditions or time points used to calculate
RNA velocity separately, default: NULL}

\item{group_column}{A string specifying the name of the metadata column in the h5ad object that contains groups.}
}
\description{
This function uses the library(anndataR) and saves a Seurat object to H5AD file. It extract and add
unspliced and spliced matrices from a list of loom files if provided.
}
