% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat_functions.R
\name{side_by_side_path_compr}
\alias{side_by_side_path_compr}
\title{Take as input a CellChat object and output the side-by-side comparison
of a pathway's signaling strength in chord diagram.}
\usage{
side_by_side_path_compr(X, object_list, cond_in_compare, pathway)
}
\arguments{
\item{X}{a merged CellChat object with two biological conditions.}

\item{object_list}{a list of CellChat object before the merge.}

\item{cond_in_compare}{a vector of condition names in comparison.}

\item{pathway}{a pathway of interest.}
}
\description{
Take as input a CellChat object and output the side-by-side comparison
of a pathway's signaling strength in chord diagram.
}
