% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scProportionTest_functions.R
\name{stat_res}
\alias{stat_res}
\title{Generate table of stat results for each comparision}
\usage{
stat_res(prop_test.i, comparisons_condition, dir_scproportion, i)
}
\arguments{
\item{prop_test.i}{scProportion object}

\item{comparisons_condition}{table of comparisions}

\item{dir_scproportion}{directory to save the results table}

\item{i}{index of the comparison condition}
}
\value{
NULL saves comparison figures in the specified directory.
}
\description{
Generate table of stat results for each comparision
}
