% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-basic.R
\name{summarize_dataset}
\alias{summarize_dataset}
\title{Summarize dataset structure}
\usage{
summarize_dataset(obj, outdir = NULL)
}
\arguments{
\item{obj}{A \code{scdown_obj}.}

\item{outdir}{Optional output directory.}
}
\value{
A list with summary tables.
}
\description{
Summarize dataset structure
}
