% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_signature}
\alias{test_signature}
\title{Test signature differences across groups}
\usage{
test_signature(obj, signatures, signature_sets = NULL, outdir = NULL)
}
\arguments{
\item{obj}{A \code{scdown_obj}.}

\item{signatures}{Character vector of signature names or a named list of signatures.}

\item{signature_sets}{Optional named list of custom signatures.}

\item{outdir}{Optional output directory.}
}
\value{
A list with signature test tables and plot paths.
}
\description{
Test signature differences across groups
}
