% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat_functions.R
\name{top_pathways}
\alias{top_pathways}
\title{Take as input(s) one or two CellChat object(s) and output 
pathways with highest overeall communication probabilities.}
\usage{
top_pathways(X1, X2 = NULL, top_n = 10)
}
\arguments{
\item{X1}{a CellChat object.}

\item{X2}{a CellChat object.}

\item{top_n}{the number of pathways.}
}
\value{
the top "top_n" pathways in a vector.
}
\description{
Take as input(s) one or two CellChat object(s) and output 
pathways with highest overeall communication probabilities.
}
