% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{get_signature}
\alias{get_signature}
\title{Get one signature}
\usage{
get_signature(name, obj = NULL)
}
\arguments{
\item{name}{Signature name.}

\item{obj}{Optional \code{scdown_obj} with custom signatures.}
}
\value{
Character vector of genes.
}
\description{
Get one signature
}
