% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{h5adToSeurat}
\alias{h5adToSeurat}
\title{Read in a h5ad file and convert to a Seurat object.}
\usage{
h5adToSeurat(h5ad_file, annotation_column = NULL, do_normalization = TRUE)
}
\arguments{
\item{h5ad_file}{character string of full path to the h5ad file.}

\item{annotation_column}{character variable specifying the metdata column name of cell type annotations.}

\item{do_normalization}{boolean variable specifying whether to perform Seurat normalization after conversion}
}
\value{
a Seurat object.
}
\description{
This function uses the library(anndataR) to convert the raw counts, metadata,
lower dimensional embeddings (pca and umap), and spliced and unspliced counts if available from a h5ad 
file to a Seurat object.
}
