% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{h5adToSeurat}
\alias{h5adToSeurat}
\title{Read in a h5ad file and transfer the AnnData object into a Seurat object.}
\usage{
h5adToSeurat(h5ad_file, annotation_column)
}
\arguments{
\item{h5ad_file}{character string of full path to the h5ad file.}
}
\value{
a Seurat object.
}
\description{
This function uses the library(anndata) to transfer the normalized counts, metadata,
and lower dimensional embeddings (pca and umap) from a h5ad file into a Seurat object.
More data can be transferred between the two formats, and this function can be expanded
to accommodate that accordingly.
}
