% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat_functions.R
\name{run_cellchatV2}
\alias{run_cellchatV2}
\title{Run CellChat}
\usage{
run_cellchatV2(
  dir_cellchat,
  seurat_obj,
  celltype_col,
  celltypes = "ALL",
  species,
  condition_col = NULL,
  conditions_cmp = NULL
)
}
\arguments{
\item{dir_cellchat}{Folder path for CellChat results.}

\item{seurat_obj}{Seurat object which contains the counts and}

\item{conditions_cmp}{}
}
\description{
run CellChat analysis and generate figures, tables and rds files.
}
