% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_monocle3.R
\name{run_monocle3}
\alias{run_monocle3}
\title{Function to run the monocle3 pipeline}
\usage{
run_monocle3(
  seurat,
  species,
  nDim = 30,
  conditions = NULL,
  condition_metadata = NULL,
  transferUMAP = TRUE,
  rootNode_method = "potency",
  rootNodes = NULL,
  timepoint = NULL,
  timePoint_metadata = NULL,
  batch_metadata = NULL,
  celltype_groups = NULL,
  top_genes = 10,
  deg_method = "quasipoisson",
  metadata_deg_model = NULL,
  graph_test = FALSE,
  cores = 1,
  outputDir = "."
)
}
\arguments{
\item{seurat}{Seurat object containing the scRNASeq data (Required)}

\item{species}{A character variable specifying the species (Required, either human or mouse)}

\item{nDim}{A numeric variable specifiying the PCA dimensions to be used for processing}

\item{conditions}{A character vector specifying the different conditions}

\item{condition_metadata}{A character variable specifying the metdata name contating conditions in the seurat object}

\item{transferUMAP}{A boolean variable specifying whether to transfer UMAP coordinates or not from Seurat object}

\item{rootNode_method}{A character variable specifying the method to identify root node (Either potency or rootNodes)}

\item{rootNodes}{A character vector specifying the principal point on the trajectory, in the format of "Y_<node number>".}

\item{timepoint}{A character vector specifying the category inside a metadata to be used to automcatically decide the root node.}

\item{timePoint_metadata}{A character vector specifying the metadata to be used to automcatically decide the root node.}

\item{batch_metadata}{A character vector specifying the name of metadata column with batch information to perform batch correction}

\item{celltype_groups}{A list of character vectors that represents a group of cell types to be subsetted and constructed trajectory by monocle}

\item{top_genes}{A numeric variable to set the number of top significant genes to visualize from DEG results}

\item{deg_method}{A character vector specifying the distribution used for regression method}

\item{metadata_deg_model}{A character vector specifying the model used for regression analysis, can be any term that exists in the metadata}

\item{graph_test}{A boolean variable to run graph auto-correlation to find differential genes along the trajectory}

\item{cores}{A numeric variable to set the number of cores to be used}

\item{outputDir}{A character vector specifying the output directory}
}
\value{
A list of monocle cell dataset objects
}
\description{
This function runs the moncole3 workflow to infer the development trajectory of the cells using the
single-cell RNASeq data. In this function, users can infer the trajectory of the whole data as well as
a subset of data and also identify the differential genes across the trajectory as well as between the
conditions.
}
