% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scProportionTest.R
\name{run_scproportion}
\alias{run_scproportion}
\title{Function to run scProportionTest pipeline}
\usage{
run_scproportion(
  seurat_obj,
  annotation_column,
  group_column,
  comparision1 = NULL,
  comparision2 = NULL,
  output_dir = ".",
  output_format = "png",
  verbose = TRUE,
  cores = 8
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing count data and metadata.}

\item{annotation_column}{The name of the metadata column in `seurat_obj` containing cluster labels or cell type names.}

\item{group_column}{The name of the metadata column in `seurat_obj` that contains sample identifiers.}

\item{comparision1}{Optional: name of the first group for comparison (default NULL, which compares all pairs).}

\item{comparision2}{Optional: name of the second group for comparison (default NULL, which compares all pairs).}

\item{output_dir}{Directory path where output figures will be saved.}

\item{output_format}{Format of the output figure}

\item{verbose}{Print the processing stept}

\item{cores}{Numenr of the cores for parallel computation}
}
\value{
NULL Save comparison figures and results in the specified directory.
}
\description{
This function runs scProportionTest for all the pairwise comparision conditions
in parallel and generate figures and table of results.
The user can aslo chose specific conditions for comparsion rather than all the comparision.
}
