% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scProportionTest_functions.R
\name{run_scproportion}
\alias{run_scproportion}
\title{Run scProportionTest}
\usage{
run_scproportion(
  dir_scproportion,
  seurat_obj,
  cluster_col,
  sample_col,
  comparision1 = NULL,
  comparision2 = NULL,
  output.format = "png",
  verbose = TRUE,
  num_cores = detectCores() - 1
)
}
\arguments{
\item{dir_scproportion}{Directory path where output figures will be saved.}

\item{seurat_obj}{A Seurat object containing count data and metadata.}

\item{cluster_col}{The name of the metadata column in `seurat_obj` containing cluster labels or cell type names.}

\item{sample_col}{The name of the metadata column in `seurat_obj` that contains sample identifiers.}

\item{comparision1}{Optional: name of the first group for comparison (default NULL, which compares all pairs).}

\item{comparision2}{Optional: name of the second group for comparison (default NULL, which compares all pairs).}

\item{output.format}{Format of the output figure}

\item{verbose}{Print the processing stept}

\item{num_cores}{Numenr of the cores for parallel computation}
}
\value{
NULL saves comparison figures in the specified directory.
}
\description{
Run scProportionTest
}
